requirements.txt
django==2.1.3
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0